+++ plugins/graphql-kotlin-gradle-plugin/src/test/resources/templates/JUnit.mustache
@@ -1,7 +1,7 @@
 package com.example
 
 import com.example.generated.JUnitQuery
-import com.expediagroup.graphql.client.GraphQLKtorClient
+import com.expediagroup.graphql.client.ktor.GraphQLKtorClient
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
